fix(save): change type of Itinerary object to Long
- it used to be int, now it is Long because we return the id of itinerary as Long
- Also the same on db

+++ src/backend/spring/src/main/java/com/example/demo/controller/ItineraryController.java
@@ -87,7 +87,7 @@ public class ItineraryController {
     @GetMapping("/user")
     public ResponseEntity<?> getUserItineraries(@RequestParam String token) {
         if (userService.verifyToken(token)) {
-            Map<Integer, Map<String, Object>> itineraries = itineraryService.getUserItineraries(token);
+            Map<Long, Map<String, Object>> itineraries = itineraryService.getUserItineraries(token);
             return ResponseEntity.ok(itineraries);
         } else {
             Map<String, String> response = new HashMap<>();
